Update redis-store.service.ts
I'm having an issue with the Redis Store. I was exploring the source code of Foal.

The thing is: the Redis Client is a private field of RedisStore (wich it's ok). But... I'm using Azure Cache for Redis (Basic version), and sometimes (very often) the connection is interrupted.

The known solution for the interruptions, is to catch the redisClient errors, allowing the client to do the reconnection (instead of closing the app).

In order to do that, I need the possibility to access to the redisClient field.

Is there any chance to add a public getter for the redis client, in RedisStore?

In order to access to that getter, the people should cast the SessionStore to RedisStore.